project-tree: Ignore untouched files
Many times when coding is useful to hide all the files not related to
the change. This feature hides all untouched files. This is just a
WIP, since it needs something like ide_vcs_monitor_add_parents to detect
directories with changes, by now it considers all directories as changed

https://bugzilla.gnome.org/show_bug.cgi?id=791897